+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Unreleased
 
+### Bug Fixes
+* **terra-suggestion-box** prevent execution of `ngModelChange`-Callback if value changes from `undefined` to `null` or reverse.
+
+<a name="2.3.12"></a>
+# 2.3.12 (20.08.2018)
+
+### Feature
+* **terra-tag-select** add new component for tag select and deselect with ngModel support
+
+<a name="2.3.11"></a>
+# 2.3.11 (17.08.2018)
+
 ### Feature
 * **category-data.interface** && **category-detail-data.interface** extend interfaces since some properties were missing in comparision to the [rest interface](https://developers.plentymarkets.com/rest-doc/category_category/details).
 * **terra-suggestion-box** 
@@ -9,6 +21,7 @@
     * `selectedValue = null` if entered text does not match any of the `inputListBoxValues`
     * `resetComponentValue`-Method is now deprecated. Use `ngModel` instead to set the value to `null`
     * `outputValueChanged` is now deprecated. Use `ngModelChange` instead.
+* **terra-form** Added new component to generate forms dynamically from json data.
     
 ### Bug Fixes
 * **terra-button** fixed disabled state showing the buttons background color while clicked or focused